+++ Common/src/main/java/fuzs/puzzleslib/impl/client/gui/WidgetTooltipHolderImpl.java
@@ -1,7 +1,6 @@
 package fuzs.puzzleslib.impl.client.gui;
 
 import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.AbstractWidget;
 import net.minecraft.client.gui.components.Tooltip;
@@ -33,9 +32,9 @@ public final class WidgetTooltipHolderImpl extends WidgetTooltipHolder {
     private final Supplier<List<? extends FormattedText>> tooltipLinesSupplier;
 
     public WidgetTooltipHolderImpl(AbstractWidget abstractWidget, TooltipBuilderImpl builder) {
-        Preconditions.checkState(!builder.tooltipLines.isEmpty() || builder.tooltipLinesSupplier != null, "lines is empty");
         this.abstractWidget = abstractWidget;
-        this.tooltipLines = builder.tooltipLinesSupplier != null ? Collections.emptyList() : ImmutableList.copyOf(builder.tooltipLines);
+        this.tooltipLines =
+                builder.tooltipLinesSupplier != null ? Collections.emptyList() : List.copyOf(builder.tooltipLines);
         this.tooltipLineProcessor = builder.tooltipLineProcessor;
         this.tooltipPositionerFactory = builder.tooltipPositionerFactory;
         this.tooltipLinesSupplier = builder.tooltipLinesSupplier;
@@ -62,12 +61,18 @@ public final class WidgetTooltipHolderImpl extends WidgetTooltipHolder {
 
     @Override
     public void setDelay(Duration delay) {
-        // NO-OP
+        // our tooltip implementation is immutable, when something tries to change it reset to vanilla
+        WidgetTooltipHolder holder = new WidgetTooltipHolder();
+        this.abstractWidget.tooltip = holder;
+        holder.setDelay(delay);
     }
 
     @Override
     public void set(@Nullable Tooltip tooltip) {
-        // NO-OP
+        // our tooltip implementation is immutable, when something tries to change it reset to vanilla
+        WidgetTooltipHolder holder = new WidgetTooltipHolder();
+        this.abstractWidget.tooltip = holder;
+        holder.set(tooltip);
     }
 
     @NotNull
@@ -81,20 +86,18 @@ public final class WidgetTooltipHolderImpl extends WidgetTooltipHolder {
     @Override
     public void refreshTooltipForNextRenderPass(boolean hovering, boolean focused, ScreenRectangle screenRectangle) {
         this.refreshLines(this.getLinesForNextRenderPass());
-        Preconditions.checkState(!this.tooltipLines.isEmpty(), "lines is empty");
-        super.refreshTooltipForNextRenderPass(hovering, focused, screenRectangle);
-    }
-
-    private List<? extends FormattedText> getLinesForNextRenderPass() {
-        if (this.tooltipLinesSupplier != null) {
-            return this.tooltipLinesSupplier.get();
-        } else {
-            return Collections.emptyList();
+        if (!this.tooltipLines.isEmpty()) {
+            super.refreshTooltipForNextRenderPass(hovering, focused, screenRectangle);
         }
     }
 
-    private void refreshLines(List<? extends FormattedText> lines) {
-        if (!lines.isEmpty() && !Objects.equals(lines, this.tooltipLines)) {
+    @Nullable
+    private List<? extends FormattedText> getLinesForNextRenderPass() {
+        return this.tooltipLinesSupplier != null ? this.tooltipLinesSupplier.get() : null;
+    }
+
+    private void refreshLines(@Nullable List<? extends FormattedText> lines) {
+        if (lines != null && !Objects.equals(lines, this.tooltipLines)) {
             this.tooltipLines = lines;
             this.get().cachedTooltip = null;
         }